@@ -1,9 +1,9 @@
 name: mnemonic
 description: Package for generating four letter Mnemonics
-version: 0.1.0
+version: 0.2.0
 
 environment:
-  sdk: '>=2.12.0 <3.0.0'
+  sdk: '>=3.5.0 <4.0.0'
 
 dependencies:
   flutter:
