@@ -9,12 +9,11 @@ class MnemonicDisplay extends StatelessWidget {
     Key? key,
     required this.mnemonic,
     this.color,
-  })  : assert(mnemonic != null),
-        super(key: key);
+  }) : super(key: key);
 
   @override
   Widget build(BuildContext context) {
-    final textColor = color ?? Theme.of(context).textTheme.headline6!.color!;
+    final textColor = color ?? Theme.of(context).textTheme.titleLarge!.color!;
 
     return Column(
       mainAxisSize: MainAxisSize.min,
@@ -28,7 +27,7 @@ class MnemonicDisplay extends StatelessWidget {
               mnemonic.acronym,
               style: Theme.of(context)
                   .textTheme
-                  .headline4!
+                  .headlineMedium!
                   .copyWith(color: textColor),
             ),
           ),
